Add official assets, enable file attach beavior
Replace placeholder assets with official ones, change file attach behavior to use absolute paths (so download URLs are able to stay the same, since we're using the front-matter array as "attached files", not inline files of any kind). This includes a custom build-time validation to ensure files exist before pushing (to hopefully reduce bugs). Keep in mind that this does not run on live-reload but will run on first build